Add the EnvLoader
This class uses the ParamsConfig to fetch any environment-configured
values from the supplied env (which is typically ENV, in real usage).

It does some basic checking and casting on those env strings (intended
to match the behavior of environment_helpers, generally), and then
returns a Hash containing all of the env *supplied* parameters, with
their param-name as the key and their type-cast ENV-value as the value.