support multi-valued parameters from environment
This made the class complex enough that it was worth factorying out an
EnvReader, for reading/parsing/casting a _single_ param's data from ENV.

In particular, to accomodate multi-valued parameters through ENV, we
allow an ENV['FOO'] to be accompanied by (or replaced by) ENV['FOO_0']
through ENV['FOO_9'] - the suffix-less variable is used first, followed
by all 10 of the others in order. (Though you're not really intended to
use them _together_, I allowed it because it's sensible to be able to
use ENV['FOO'] even if FOO is multi-valued, since that most closely
mimics standard CLI param behavior.